+++ packages/server/core/src/types/config/server.ts
@@ -31,6 +31,7 @@ export interface ServerUserConfig {
   watchOptions?: WatchOptions;
   compiler?: 'babel' | 'typescript';
   enableFrameworkExt?: boolean;
+  worker?: boolean;
 }
 
 export type ServerNormalizedConfig = ServerUserConfig;