+++ packages/server/plugin-worker/modern.config.js
@@ -1,0 +1,5 @@
+const { legacyPresets } = require('@scripts/build');
+
+module.exports = {
+  buildConfig: legacyPresets.UNIVERSAL_JS,
+};