+++ packages/builder/plugin-node-polyfill/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @modern-js/builder-plugin-node-polyfill
 
+## 2.4.0
+
 ## 2.3.0
 
 ## 2.2.0