+++ packages/cli/babel-preset-app/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @modern-js/babel-preset-app
 
+## 2.4.0
+
+### Patch Changes
+
+- Updated dependencies [98a2733]
+- Updated dependencies [8c2db5f]
+  - @modern-js/utils@2.4.0
+  - @modern-js/babel-preset-base@2.4.0
+
 ## 2.3.0
 
 ### Patch Changes